added pdfRenderer in a new .tsx file
