+++ components/brave_rewards/resources/brave_rewards_static_resources.grd
